+++ ilogtail-controller/controller/base_setting.json
@@ -1,0 +1,4 @@
+{
+    "store_mode":"1",
+    "run_mode":"0"
+}
