+++ ilogtail-controller/controller/dataDB/CURRENT
@@ -1,0 +1,1 @@
+MANIFEST-000003
