--- ilogtail-controller/server/base_setting.json
@@ -1,4 +1,0 @@
-{
-    "store_mode":"1",
-    "run_mode":"0"
-}
