--- ilogtail-controller/server/tool/json.go
@@ -1,19 +1,0 @@
-package tool
-
-import (
-	"encoding/json"
-	"os"
-)
-
-func ReadJson(filePath string, ptr interface{}) {
-	file, err := os.Open(filePath)
-	if err != nil {
-		panic(err)
-	}
-	defer file.Close()
-	decoder := json.NewDecoder(file)
-	err = decoder.Decode(&ptr)
-	if err != nil {
-		panic(err)
-	}
-}
